backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
supabase==2.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pandas==2.1.4
openpyxl==3.1.2
numpy==1.26.3
python-dateutil==2.8.2
httpx==0.26.0
websockets==12.0
redis==5.0.1
celery==5.3.4
loguru==0.7.2
python-dotenv==1.0.0
aiofiles==23.2.1
motor==3.3.2
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
flake8==7.0.0
isort==5.13.2
